M4: Split WasteSearchTableExpandContent mega-test - Extract row expansion display
Reduced main test file from 599 lines to 482 lines (117 line reduction).
Created row-expansion-data-display.unit.test.tsx with 177 lines (9 tests).

Extracted test group:
- row expansion data display (9 tests)

All 1421 tests passing. Test file count now 122.
